% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ACTR.R
\docType{package}
\name{ACTR}
\alias{ACTR}
\alias{ACTR-package}
\title{ACTR: package to develop method to generate EDR10 from ED50}
\details{
The method contains X steps:

\itemize{
\item Step 1 :
\item Step 2 :
\item etc
}
}

