% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plottings.R
\name{actR_ggtheme}
\alias{actR_ggtheme}
\title{Define a plot theme for actR}
\usage{
actR_ggtheme(font.family = "Helvetica", ...)
}
\arguments{
\item{font.family}{Specify a font family to use for the theme (default = "Helvetica")}

\item{...}{parameters to pass to theme function}
}
\description{
Use this to define a theme across actR
}
\concept{plot help}
