% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/synth.R
\name{ar1noise}
\alias{ar1noise}
\title{AR(1) noise}
\usage{
ar1noise(lngth, g, sigma)
}
\arguments{
\item{lngth}{number of samples}

\item{g}{lab-1 autocorrelation (short-term memory parameter)}

\item{sigma}{standard deviation of the Gaussian innovations}
}
\value{
lngth samples of the process (vector)
}
\description{
samples from AR(1) process
}
