% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assert_distinct.R
\name{assert_distinct}
\alias{assert_distinct}
\title{assert_distinct}
\usage{
assert_distinct(data, ...)
}
\arguments{
\item{...}{}
}
\value{

}
\description{
assert_distinct
}
\examples{
assert_distinct(my.data.frame, key1, key2, key3)
}
