% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bestModel_mu.R
\name{bestModel_mu}
\alias{bestModel_mu}
\title{Selecting the best model, by bootstrap and Cross Validation,  to predict wMuC_lg}
\usage{
bestModel_mu(errMod_Mu.list)
}
\arguments{
\item{errMod_Mu.list}{A list containing the error of prediction of wMuC_lg for each model
and each bootstrap samples}
}
\value{
The function return the formula of the best selected model
}
\description{
6 linear models are compared using Cross Validation on the basis of
the least error of prediction
}
\details{
The 6 linear models are :

mod1  : wMuC_lg ~ 1

mod2  : wMuC_lg ~ wMuA_lg

mod3  : wMuC_lg ~ wMuA_lg + wSigmaA_lg

mod4  : wMuC_lg ~ wMuA_lg + I(wMuA_lg^2)

mod5  : wMuC_lg ~ wMuA_lg + I(wMuA_lg^2) + wSigmaA_lg

mod6  : wMuC_lg ~ wMuA_lg + I(wMuA_lg^2) + wSigmaA_lg + I(wSigmaA_lg^2)
}
\examples{
data(cipr)
ciprKP <- subset_data(cipr, Class,6)
PAC.bt <- lapply(ciprKP.bt,est_PAC,Class)
errMod_mu <- lapply(PAC.bt,msep_mu)
errMod_sigma <- lapply(PAC.bt,msep_sigma)
bm_mu <- bestModel_mu(errMod_mu)
}

