% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boot_data.R
\name{boot_data}
\alias{boot_data}
\title{Build bootstrap sample of data according a grouping variable (i.e. Class or Phylum)}
\usage{
boot_data(dat, var_grp, B)
}
\arguments{
\item{var_grp}{A grouping variable}

\item{B}{The number of bootstrap replicates}

\item{dataset}{A dataframe}
}
\value{
The function returns a list containing a bootstrap sample in each element
}
\description{
B bootstrap samples of an original dataset are build with replacement using strata as
 variable group (i.e.Class or Phylum ) and DoseType (Acute / Chronic)
}
\examples{
data(cipr)
new <- subset_data(cipr, Class,6)
new.boot<-boot_data(ciprKP,Class,200)
}

