% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial.R
\name{calculateMultiTestSignificance}
\alias{calculateMultiTestSignificance}
\title{Title}
\usage{
calculateMultiTestSignificance(events, weights = NA, n.ens = 1000)
}
\arguments{
\item{n.ens}{}
}
\description{
Title
}
