% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_levels.R
\name{check_all}
\alias{check_all}
\title{Print levels of different variables in order to check data are ready for
applying ACT method}
\usage{
check_all(dataset)
}
\arguments{
\item{dataset}{A dataframe containing a variable named "Class"}
}
\description{
cf help page of check_no_mutation , check_dosetype, check_class
and check_species_byclass functions
}
\details{
Are printed :

Levels of Umbrella effect (no mutation expected)

Levels of dose type (are printed)only chronic and acute expected)

Levels of class  (strictly same orthography needed)

This function calls following functions:

check_no_mutation,

check_dosetype

check_class

check_species_byclass

 Outputs are printed in the "check_all_output.txt " file
}
\examples{
check_all(cipr)
}

