% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_levels.R
\name{check_class}
\alias{check_class}
\title{Print all the levels (modalities) of Class}
\usage{
check_class(dataset)
}
\arguments{
\item{dataset}{A dataframe containing a variable named "Class"}
}
\description{
Aim : check each Class'name is writtten with the
same orthography (needed for the ACT method)
}
\details{
Outputs are printed in file
}
\examples{
check_class(cipr)
}

