% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_levels.R
\name{check_dosetype}
\alias{check_dosetype}
\title{Print all the levels (modalities)  of Dose Type}
\usage{
check_dosetype(dataset)
}
\arguments{
\item{dataset}{A dataframe containing a variable named "DoseType"}
}
\description{
Aim : check there are only CHRONIC and ACUTE dose types in the datset
}
\details{
Outputs are printed in file
}
\examples{
check_dosetype(cipr)
}

