% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_levels.R
\name{check_no_mutation}
\alias{check_no_mutation}
\title{Print all the levels (modalities) of Umbrella Effects}
\usage{
check_no_mutation(dataset)
}
\arguments{
\item{dataset}{A dataframe containing a variable named "Umbrella"}
}
\description{
Aim : check that there i no mutation effect in the dataset
}
\details{
Outputs are printed in file
}
\examples{
check_no_mutation(cipr)
}

