% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_levels.R
\name{check_species_byclass}
\alias{check_species_byclass}
\title{Print all the levels (modalities) of Species for each Class of for each Dose Type}
\usage{
check_species_byclass(dataset)
}
\arguments{
\item{dataset}{A dataframe containing DoseType, Class and SpeciesComp variables}
}
\description{
Aim :  check that species names are always writtten with the
same orthography (needed for the ACT method)
}
\details{
Outputs are printed in file
}
\examples{
check_species_byclass(cipr)
}

