% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coef_reg.R
\name{coef_reg}
\alias{coef_reg}
\title{Coefficient of best models applied on observed data}
\usage{
coef_reg(PAC_obs, form_mu, form_sigma)
}
\arguments{
\item{PAC_obs}{dataframe containing the observed  acute and chronic weighted parameters}

\item{form_mu}{Formula of the best model selected to predict muC_act}

\item{form_sigma}{Formula of the best model selected to predict SigmaC_act}
}
\value{
a list with the regression coefficients of the best model applied on observed data
}
\description{
Coefficient of best models applied on observed data
}
\examples{
library(ACTR)
data(cipr)
ciprKP <- subset_data(cipr, Class,6)
PAC_obs <- est_PAC(ciprKP,Class)
coef_Mod <- coef_reg (PAC_obs,bm_mu, bm_sigma)
}

