% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/excursion.R
\name{detectExcursionCore}
\alias{detectExcursionCore}
\title{Detect excursion - core functionality}
\usage{
detectExcursionCore(age, vals, event.start, event.end, sig.num, n.consecutive)
}
\arguments{
\item{age}{a year or age vector}

\item{vals}{a vector of values the same length as age}

\item{event.start}{the start of the event window}

\item{event.end}{the end of the event window}

\item{sig.num}{how many standard deviations required outside the reference windows must be exceeded for this to be considered an excursion? (default = 2)}

\item{n.consecutive}{how many consecutive points are required for this to be considered an excursion? (default = 2)}
}
\value{
a tibble of results
}
\description{
Detect excursion - core functionality
}
