% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/excursion.R
\name{detectExcursionMcSim}
\alias{detectExcursionMcSim}
\title{Detect excursions in synthetic datasets that mimic a real on}
\usage{
detectExcursionMcSim(
  age,
  vals,
  event.yr,
  event.window,
  ref.window,
  n.ens = 100,
  n.consecutive = 2,
  sig.num = 2
)
}
\arguments{
\item{age}{a year or age vector}

\item{vals}{a vector of values the same length as age}

\item{event.yr}{the center of the proposed excursion window}

\item{event.window}{the width of the proposed excursion window}

\item{ref.window}{how many years to use as a reference before and after the event window}

\item{n.ens}{The number of ensembles to simulate (default = 100)}

\item{n.consecutive}{how many consecutive points are required for this to be considered an excursion? (default = 2)}

\item{sig.num}{how many standard deviations required outside the reference windows must be exceeded for this to be considered an excursion? (default = 2)}
}
\value{
a tibble that reports the positivity rate in the synthetics
}
\description{
Detect excursions in synthetic datasets that mimic a real on
}
