% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/excursion.R
\name{detectExcursionSlidingWindow}
\alias{detectExcursionSlidingWindow}
\title{Detect an excursion in many timeseries}
\usage{
detectExcursionSlidingWindow(
  ltt = NA,
  time = NA,
  vals = NA,
  time.variable.name = NA,
  vals.variable.name = NA,
  time.units = NA,
  vals.units = NA,
  dataset.name = NA,
  event.yr.vec = NA,
  event.step = NA,
  ref.window,
  ...
)
}
\arguments{
\item{ltt}{A LiPD-timeseries-tibble, a tibble or data.frame that has the variable(s) of interest, a time variable (age, year or time) along with their metadata, aranged in rows. If ltt = NA, then one in is created from other inputs}

\item{time}{if ltt is not provided, input a vector or matrix of time (year or age) data. If it's a multicolumn matrix, the columns are time-ensemble members}

\item{vals}{if ltt is not provided, input a vector or matrix of paleoData. If it's a multicolumn matrix, the columns are value-ensemble members}

\item{time.variable.name}{If ltt is not provided, specify the name of the time variable (typically 'age' or 'year')}

\item{vals.variable.name}{If ltt is not provided, specify the name of the paleo variable (e.g., 'd18O' or 'temperature'). Alternatively, if ltt is provided with more rows than expected, this term is used to attempt to select the correct row.}

\item{time.units}{If ltt is not provided, specify the units the time variable (typically 'yr BP' or 'CE')}

\item{vals.units}{If ltt is not provided, specify the units the paleo variable (e.g. 'permil' or 'degrees C')}

\item{dataset.name}{If ltt is not provided, specify the dataset name}

\item{event.yr.vec}{A vector of times at the center of event years to test.}

\item{event.step}{If event.yr.vec = NA, then event.step will build one with this spacing.}

\item{ref.window}{width (in time units) of the reference windows}

\item{...}{pass additional arguments to detectExcursion()}
}
\value{
a tibble that describes the positive and negative excursion results
}
\description{
Determines whether an excursion event has occurred within the specified event window for a lipd-ts-tibble of timeseries. Excursion events are defined as n.consecutive values within the event window that are more extreme than the avg +/- sig.num standard deviations of the reference windows.
}
\references{
Morrill
}
\author{
Hannah Kolus

Nick McKay
}
