% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ruggieri.R
\name{detectKinkBayesCore}
\alias{detectKinkBayesCore}
\title{Detects changes in trend (a kink) in series Y based on design matrix X}
\usage{
detectKinkBayesCore(
  time,
  vals,
  gaussianize = TRUE,
  k_max = 10,
  d_min = 5,
  k_0 = c(0.001, 0.1),
  v_0 = 1,
  sig_0 = sd(vals),
  num_samp = 1000
)
}
\arguments{
\item{k_max}{maximum number of change points allowed}

\item{d_min}{minimum distance between adjacent change points}

\item{k_0}{hyper parameter for the prior on the regression coefficients}

\item{v_0}{prior location parameter of the Scaled − Inverse $\chi^2$ distribution for \sigma}

\item{sig_0}{prior scale parameter of the Scaled − Inverse $\chi^2$ distribution for \sigma}

\item{num_samp}{number of sampled solutions}

\item{Y}{a vector of length N}

\item{X}{design matrix,  a matrix of dimensions N x m, where m is the number of predictors}

\item{t}{time axis common to Y and X (is it needed at all?)}
}
\value{
a list containing k, \beta, chgpt_loc, samp_holder, model, size, year, and Y
}
\description{
Detects changes in trend (a kink) in series Y based on design matrix X
}
