% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/excursion.R
\name{detectMultipleExcursions}
\alias{detectMultipleExcursions}
\title{Detect an excursion in many timeseries}
\usage{
detectMultipleExcursions(
  ltt = NA,
  n.ens = 100,
  surrogate.method = "isospectral",
  null.hypothesis.n = 100,
  event.yr,
  event.window,
  ref.window,
  sig.num = 2,
  n.consecutive = 2,
  exc.type = "either",
  min.vals = 8,
  na.rm = TRUE,
  simulate.time.uncertainty = FALSE,
  simulate.paleo.uncertainty = FALSE,
  seed = as.integer(Sys.time())
)
}
\arguments{
\item{ltt}{A LiPD-timeseries-tibble, a tibble or data.frame that has the variable(s) of interest, a time variable (age, year or time) along with their metadata, aranged in rows. If ltt = NA, then one in is created from other inputs}

\item{n.ens}{How many ensembles to use for error propagation? (default = 100)}

\item{surrogate.method}{What method to use to generage surrogate data for hypothesis testing? Options include: \itemize{
\item 'isospectral': (Default) Following Ebisuzaki (1997), generate surrogates by scrambling the phases of the data while preserving their power spectrum. This uses the To generate these "isospectral" surrogates. Uses the rEDM::make_surrogate_data() or rEDM::SurrogateData() function depending on version
\item 'isopersistent':  Generates surrogates by simulating from an autoregressive process of order 1 (AR(1)), which has been fit to the data. Uses the ens::createSyntheticTimeseries() function
\item 'shuffle': Randomly shuffles the data to create surrogates. Uses the rEDM::make_surrogate_data() or rEDM::SurrogateData() function depending on version
}}

\item{null.hypothesis.n}{How many simulations to run for null hypothesis testing (default = 100)}

\item{event.yr}{time at the center of the excursion window}

\item{event.window}{width (in time units) of the excursion window}

\item{ref.window}{width (in time units) of the reference windows}

\item{sig.num}{how many standard deviations required outside the reference windows must be exceeded for this to be considered an excursion? (default = 2)}

\item{n.consecutive}{how many consecutive points are required for this to be considered an excursion? (default = 2)}

\item{exc.type}{Type of excursion to look for. "positive", "negative", "either" or "both" (default = "either")}

\item{min.vals}{Minimum effective sample size (adjusted by autocorrelation) required in reference and event windows (default = 4)}

\item{na.rm}{Remove NAs? (default = TRUE)}

\item{simulate.time.uncertainty}{TRUE or FALSE. If an ensemble is not included, do you want to simulate time ensembles (default = TRUE)}

\item{simulate.paleo.uncertainty}{TRUE or FALSE. If an ensemble is not included, do you want to simulate paleo ensembles (default = TRUE)}

\item{seed}{Set a seed for reproducibility. By default it will use current time meaning it will not be reproducible.}
}
\value{
a tibble that describes the positive and negative excursion results
}
\description{
Determines whether an excursion event has occurred within the specified event window for a lipd-ts-tibble of timeseries. Excursion events are defined as n.consecutive values within the event window that are more extreme than the avg +/- sig.num standard deviations of the reference windows.
}
\references{
Morrill
}
\author{
Hannah Kolus

Nick McKay
}
