% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial.R
\name{distanceGrid}
\alias{distanceGrid}
\title{Distance from grid cells to TS objects}
\usage{
distanceGrid(
  wts = NULL,
  lon.range = c(-179.5, 179.5),
  lat.range = c(-89.5, 89.5),
  grid.resolution = 1
)
}
\arguments{
\item{wts}{a lipd TS object}

\item{grid.resolution}{grid cell size resolution, eg. 1 deg x 1 deg}

\item{radius.group}{radius of filter for each grid cell in km}
}
\value{
distances
}
\description{
Distance from grid cells to TS objects
}
