% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/est_PAC.R
\name{est_PAC}
\alias{est_PAC}
\title{Estimation of weighted means and standard deviation of EDR10 and ED50 samples
(in log scale)}
\usage{
est_PAC(dat, var_grp)
}
\arguments{
\item{var_grp}{A group variable telling how counts have to be done (ie.by Class)}

\item{dataset}{A dataframe}

\item{n_min}{A integer telling the minimum number of EDR10 and ED50 desired}
}
\value{
a data frame with the weighted Mu_A, Mu_C, Sigma_A and Sigma_C
}
\description{
Estimation of weighted means and standard deviation of EDR10 and ED50 samples
(in log scale)
}
\details{
The function estimates weighted mean and standard deviation of EDR10 and ED50
samples. Weight ar 1/(n_i * n_s), where n_i is the number of ED mesaured on
the same species and n_s is the numbre of species in the modality of var_grp
(ie. Class).
More details into "Acute to chronic species sensivity distribution
extrapolation,  2004 (Duboudin et al, Ennv Tox & Chem , vol 23 n 7, p1774 - 1785)
}
\examples{
library(ACTR)
data(cipr)
ciprKP <- subset_data(cipr, Class,6)
myPAC <- est_PAC (ciprKP,Class)
}

