% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/est_para.R
\name{est_para}
\alias{est_para}
\title{Return wMu and wSigma of ED contained into a data frame}
\usage{
est_para(dat)
}
\arguments{
\item{dat}{A data frame containing the Ed values, SpeciesComp and var-group (i.e. Class)}
}
\value{
A data frame with the estimated parameters
}
\description{
Return wMu and wSigma of ED contained into a data frame
}
\examples{
est_para(edr10_act)
lapply(edr10_obs_list, est_para)
}

