% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/est_PC_act.R
\name{est_PC_act}
\alias{est_PC_act}
\title{Prediction of wMuC_act and wSigmaC_act, i.e on edr10_act}
\usage{
est_PC_act(edr10_act, var_grp)
}
\arguments{
\item{edr10_act}{A dataframe containing the ACT edr10}

\item{var_group}{A variable of regroupement as "Class" or "Phylum"}
}
\value{
a dataframe
}
\description{
Prediction of wMuC_act and wSigmaC_act, i.e on edr10_act
}
\examples{
library(ACTR)
data(cipr)
ciprKP <- subset_data(cipr, Class,6)
PAC_obs <- est_PAC(ciprKP,Class)
PC_pred <- predict_PC(PAC_obs,bm_mu, bm_sigma)
PAC_gen <- cbind(PA_obs,PC_pred[,-1] )
edr10_act <- gen_edr10(PAC_gen,ciprKP.ac)
PC_act <- est_PC_act(edr10_act,Class)
}

