% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_edr10.R
\name{gen_edr10}
\alias{gen_edr10}
\title{Genertae edr10_act from ED50 (acute data) and regression coefficients}
\usage{
gen_edr10(PAC_gen, acute_df)
}
\arguments{
\item{acute_df}{A dataframe containing the ED50}

\item{paracACT}{A dataframe containing mu_C_act and sigma_C_acy}
}
\value{
a data frame containing the predicted EDR10 (=EDR10_act)
}
\description{
ED50 are standardized (substraction mu_A and dividing by sigma_ A)
and then scaled into chronic data (ED10) adding muC_act and multiplying by
sigma_C_act
}
\examples{
library(ACTR)
data(cipr)
ciprKP <- subset_data(cipr, Class,6)
PAC_obs <- est_PAC(ciprKP,Class)
PC_pred <- predict_PC(PAC_obs,bm_mu, bm_sigma)
PAC_gen <- cbind(PA_obs,PC_pred[,-1] )
edr10_act <- gen_edr10(PAC_gen,ciprKP.ac)
}

