% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{geob7926_2}
\alias{geob7926_2}
\title{Kim et al. GeoB7926-2 Temperature record, with age and paleo ensembles}
\format{
A LiPD-TS-tibble with 13 LiPD variable columns, age and temperature ensembles
}
\source{
\url{https://github.com/nickmckay/Temperature12k/}
}
\usage{
geob7926_2
}
\description{
Kim et al. GeoB7926-2 Temperature record, with age and paleo ensembles
}
\keyword{datasets}
