% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/excursion.R
\name{getAllExcursionTimes}
\alias{getAllExcursionTimes}
\title{Get all excursion times from an output}
\usage{
getAllExcursionTimes(x, exc.type = "Either")
}
\arguments{
\item{x}{an excursion object}

\item{exc.type}{can be Either, Above or Below}
}
\value{
a vector of ages that were identified to as excursions
}
\description{
Get all excursion times from an output
}
