% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/excursion.R
\name{getAllTimes}
\alias{getAllTimes}
\title{get the times from a single eventDetection object}
\usage{
getAllTimes(ed, exc.type = "Either")
}
\arguments{
\item{ed}{event detection object}

\item{exc.type}{excursion type}
}
\value{
return an unlisted vector of times
}
\description{
get the times from a single eventDetection object
}
