% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial.R
\name{getNulls}
\alias{getNulls}
\title{Title}
\usage{
getNulls(allNulls, nrows, n.ens, weights = NA)
}
\arguments{
\item{allNulls}{all the null data}

\item{nrows}{number of rows}

\item{n.ens}{number of ensembles}

\item{weights}{a vector to optionally weight the nulls}
}
\value{
a vector of null outputs
}
\description{
Title
}
