% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{great_basin}
\alias{great_basin}
\title{Salzer et al. Great basin tree ring record}
\format{
A LiPD-TS-tibble with 13 LiPD variable columns
}
\source{
\url{https://lipdverse.org/Pages2kTemperature/2_1_0/NAm-GreatBasin.Salzer.2013.html}
}
\usage{
great_basin
}
\description{
Salzer et al. Great basin tree ring record
}
\keyword{datasets}
