% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Hdr_Emp.R
\name{Hdr_Emp}
\alias{Hdr_Emp}
\title{Calculating HDR5 of observed data from log normal distribution}
\usage{
Hdr_Emp(edr10.df, th)
}
\arguments{
\item{th}{The desired probability (i.e. 0.05 for HDR5)}

\item{edr10}{A dataframe containing the edr10 for all the var_group  we want to calculate hdr th}
}
\value{
a data frame containing the hdr th calculated from empirical distribution
}
\description{
Calculating HDR5 of observed data from log normal distribution
}
\examples{
Hdr_Emp(edr10_obs,0.05)
}

