% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Hdr_LN.R
\name{Hdr_LN}
\alias{Hdr_LN}
\title{Calculating HDR th  of data from log normal distribution}
\usage{
Hdr_LN(edr10.df, th)
}
\arguments{
\item{th}{The desired probability (i.e. 0.05 for HDR5)}

\item{edr10}{A dataframe containing the edr10 for all the var_group  we want to calculate hdr th}
}
\value{
a data frame containing the hdr th calculated from empirical distribution
}
\description{
Calculating HDR th  of data from log normal distribution
}
\examples{
Hdr_LN(edr10_obs,0.05)
}

