% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Hdr_LN2.R
\name{Hdr_LN2}
\alias{Hdr_LN2}
\title{Calculating HDR5 of observed data from log normal distribution.}
\usage{
Hdr_LN2(edr10.df, th)
}
\arguments{
\item{th}{The desired probability (i.e. 0.05 for HDR5)}

\item{edr10}{A dataframe containing the edr10 for all the var_group  we want to calculate hdr th}
}
\value{
a data frame containing the hdr th calculated from empirical distribution
}
\description{
Calculating HDR5 of observed data from log normal distribution.
}
\details{
This function is necessary for IC_HdrLN2, when IC of hdr th are etsimated without caclculation weight of the bootstrap sample
}
\examples{
Hdr_LN(edr10_obs,0.05)
}

