% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/highSchoolLookup.R
\name{highSchoolLookup}
\alias{highSchoolLookup}
\title{Finding ACT High School Codes}
\usage{
highSchoolLookup(state, country = "US", city = "", name = "")
}
\arguments{
\item{state}{Should be the two character state/province abbreviation}

\item{country}{Should be the abbreviated character list}

\item{city}{Optional - name of the city where the high school is located}

\item{name}{Optional - name of the school to lookup}
}
\value{
Returns a data frame containing the response payload
}
\description{
Provides a simple function to form/request high school lookup
codes from the ACT API
}
\examples{
# Get all the high schools located in Kentucky
kyHighSchools <- actR::highSchoolLookup("US", "KY")

# Display part of the dataframe
head(kyHighSchools)


}
