% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IC_HdrEmp.R
\name{IC_HdrEmp}
\alias{IC_HdrEmp}
\title{Calculating IC 95 percent of empirical hdr using non parametric bootstrap samples}
\usage{
IC_HdrEmp(edr10.df, th, B)
}
\arguments{
\item{th}{The desired probability (i.e. 0.05 for HDR5)}

\item{B}{TNumber of bootstrap samples}

\item{edr10}{A dataframe containing the edr10 for all the var_group  we want to calculate hdr th}
}
\value{
a data frame containing the hdr calculated from empirical distribution and the 95 percent confidence interval
}
\description{
Calculating IC 95 percent of empirical hdr using non parametric bootstrap samples
}
\examples{
IC_HdrEmp(edr10_obs, 0.05,199)
}

