% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IC_HdrLN.R
\name{IC_HdrLN}
\alias{IC_HdrLN}
\title{Calculating IC 95 percent of  hdr estimated from log normal distribution, using non parametric bootstrap samples}
\usage{
IC_HdrLN(edr10.df, th, B)
}
\arguments{
\item{th}{The desired probability (i.e. 0.05 for HDR5)}

\item{B}{TNumber of bootstrap samples}

\item{edr10}{A dataframe containing the edr10 for all the var_group  we want to calculate hdr th}
}
\value{
a data frame containing the hdr calculated from log normal distribution and the 95 percent confidence interval
}
\description{
Calculating IC 95 percent of  hdr estimated from log normal distribution, using non parametric bootstrap samples
}
\examples{
hdr5_LNA <- IC_HdrLN_act(ciprKP, Class, bm_mu, bm_sigma, 0.05,1000)
}

