% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IC_HdrLN2.R
\name{IC_HdrLN2}
\alias{IC_HdrLN2}
\title{Calculating IC 95 percent of  hdr estimated from log normal distribution, using non parametric bootstrap samples and without calculating weight accordind species}
\usage{
IC_HdrLN2(edr10.df, th, B)
}
\arguments{
\item{th}{The desired probability (i.e. 0.05 for HDR5)}

\item{B}{TNumber of bootstrap samples}

\item{edr10}{A dataframe containing the edr10 for all the var_group  we want to calculate hdr th}
}
\value{
a data frame containing the hdr calculated from log normal distribution and the 95 percent confidence interval
}
\description{
Calculating IC 95 percent of  hdr estimated from log normal distribution, using non parametric bootstrap samples and without calculating weight accordind species
}
\examples{
IC_HdrLN(edr10_obs, 0.05,199)
}

