% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IC_HdrLN_act.R
\name{IC_HdrLN_act}
\alias{IC_HdrLN_act}
\title{Calculating IC of HDR (th) of edr10_act from log normal distribution}
\usage{
IC_HdrLN_act(dat, varGrp, BM_Mu, BM_Sigma, th, B)
}
\arguments{
\item{th}{The desired probability (i.e. 0.05 for HDR5)}

\item{B}{The desired number of bootstrap sample from which confidence intervals are build}

\item{data}{A dataframe with the data}

\item{bestModelMu}{A linear model}

\item{bestModelSigma}{A linear model}
}
\value{
a data frame with estimated HDR5_act and 95% confidence interval (not in log scale)
}
\description{
Number of bootstrap samples is determinated by PAC_boot
}
\examples{
hdr5_act <- HDR5_act ( ciprKP, BM_Mu, BM_Sigma,PAC_boot )
}

