% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/isValidCountry.R
\name{isValidCountry}
\alias{isValidCountry}
\title{Internal function for validating entries}
\usage{
isValidCountry(countryCode)
}
\arguments{
\item{countryCode}{is the country code value being validated}
}
\value{
Returns a boolean indicating if the country code passed was valid
}
\description{
Function used to test whether a valid country code was passed
}
