% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plottings.R
\name{kdePval}
\alias{kdePval}
\title{Use a kde to estimate pvalue}
\usage{
kdePval(nulls, real, h = NA)
}
\arguments{
\item{nulls}{a vector of null hypothesis results}

\item{real}{a single value for the real data}

\item{h}{smoothing factor}
}
\value{
a list with pvalues and kde data
}
\description{
Use a kde to estimate pvalue
}
