% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/synth.R
\name{linearRamp}
\alias{linearRamp}
\title{Linear ramp}
\usage{
linearRamp(lngth, width)
}
\arguments{
\item{lngth}{length of the time axis}

\item{width}{the width of the ramp. width  = 0 results in a step jump at the midpoint. width = lngth/2 (the largest alloweable value) results in a smooth linear ramp}
}
\value{
values of the function (between 0 and 1)
}
\description{
linearRamp()  Create a linear ramp around the midpoint of a series. Slope can be adjusted as desired to create gradual or abrupt shifts.
}
