% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make.R
\name{makeExcursion}
\alias{makeExcursion}
\title{Build a synthetic LiPD-timeseries-tibble with an excusrion}
\usage{
makeExcursion(
  time = NA,
  length = 1000,
  amplitude = 1,
  delta.amplitude = 0,
  start.time = 400,
  duration = 200,
  snr = 1,
  sample.spacing.frac = 1,
  noise.ar1 = 0.5,
  ...
)
}
\arguments{
\item{time}{time vector over which to make the excursion}

\item{length}{number of time steps in series}

\item{amplitude}{amplitude of excursion}

\item{delta.amplitude}{amplitude of change during excursion}

\item{start.time}{beginning time step of excursion}

\item{duration}{duration of excursion}

\item{snr}{signal to noise ratio}

\item{sample.spacing.frac}{if < 1, fraction of the data to retain, randomly removing others}

\item{noise.ar1}{AR1 coefficient for the noise}

\item{...}{parameters passed to prepareInput}
}
\value{
ltt
}
\description{
Build a synthetic LiPD-timeseries-tibble with an excusrion
}
