% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make.R
\name{makeShift}
\alias{makeShift}
\title{Build a synthetic LiPD-timeseries-tibble with a shift in mean}
\usage{
makeShift(length, amp, start, dur = 0, shift = 0, ...)
}
\arguments{
\item{length}{number of time steps}

\item{amp}{amplitude of shift}

\item{start}{starting time step of shift}

\item{dur}{duration of mean shift}

\item{shift}{value of time series before shift}

\item{...}{values passed to prepareInput}
}
\value{
ltt
}
\description{
Build a synthetic LiPD-timeseries-tibble with a shift in mean
}
