% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/msep_sigma.R
\name{msep_sigma}
\alias{msep_sigma}
\title{Estimating the Mean Squared Error of Prediction for Sigma parameter, of several linear models, using cross validation}
\usage{
msep_sigma(PAC.df)
}
\arguments{
\item{PAC.df}{A dataframe containing the weighted parameters in log10 scale (wMuA_lg, wMuC_lg, wSigmaA_lg, wSigmaC_lg)}
}
\value{
The function return the formula of the best selected model
}
\description{
6 linear models linear are compared using Cross Validation on the basis of
the least error of prediction (cf Details)
}
\details{
The 6 linear models are :

mod1  :  wSigmaC_lg ~ 1

mod2  :  wSigmaC_lg ~ wSigmaA_lg

mod3  : wMuC_lg ~ wMuA_lg + wSigmaA_lg

mod4  :  wSigmaC_lg ~ wSigmaA_lg + I(wSigmaA_lg^2)

mod5  : wSigmaC_lg ~ wSigmaA_lg + I(wSigmaA_lg^2) + wMuA_lg

mod6  : wSigmaC_lg ~ wSigmaA_lg + I(wSigmaA_lg^2) + wMuA_lg + I(wMuA_lg^2)
}
\examples{
library(ACTR)
ciprKP <- subset_data(cipr, Class,6)
PAC.bt <- lapply(ciprKP.bt,est_PAC,Class)
errMod_sigma <- lapply(PAC.bt,msep_sigma)
}

