% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nb_ED.R
\name{nb_ED}
\alias{nb_ED}
\title{Print the number of available EDR10 and ED50, present in
the dataset, by group variable level.}
\usage{
nb_ED(dat, var_grp, n_min)
}
\arguments{
\item{var_grp}{A group variable telling how counts have to be done (ie.by Class)}

\item{n_min}{A integer telling the minimum number of EDR10 and ED50 desired}

\item{dataset}{A dataframe}
}
\value{
the number of Outliers according the chosen coefficient
}
\description{
The function prints the count for :  1) all the modalities of the group
variable, 2) the modalities having at least 1 EDR10 or 1 ED50,
3) modalities having at least k (argument on the function) EDR10 and k ED50.
}
\details{
#'
}
\examples{
data(cipr)
a <- nb_ED(cipr, Class, 6)
nb_ED(cipr, Phylum, 6)
}

