% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nb_outliers.R
\name{nb_outliers}
\alias{nb_outliers}
\title{Return the number of outliers in the dataset taken in entry and according the coefficient.
This function is based on boxplot.stat function}
\usage{
nb_outliers(data, var_grp, coef)
}
\arguments{
\item{data}{A dataframe containing the edr10 or ed50}

\item{var_grp}{A group variable telling how counts have to be done (ie.by Class)}

\item{coef}{this determines how far the plot ‘whiskers’ extend out from the
box. If coef is positive, the whiskers extend to the most extreme data point
which is no more than coef times the length of the box away from the box.
A value of zero causes the whiskers to extend to the data extremes (and no outliers be returned).}
}
\value{
the number of Outliers according the chosen coefficient
}
\description{
Return the number of outliers in the dataset taken in entry and according the coefficient.
This function is based on boxplot.stat function
}
\examples{
 nb_outliers(ciprKP0.Ac,Class,coef)
}

