% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ngrip}
\alias{ngrip}
\title{NGRIP Ice-core d18O record}
\format{
A LiPD-TS-tibble with 3 LiPD variable columns
}
\source{
\url{https://lipdverse.org/HoloceneAbruptChange/0_10_0/NGRIP.NGRIP.2004.html}
}
\usage{
ngrip
}
\description{
NGRIP Ice-core d18O record
}
\keyword{datasets}
