% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{odp658c}
\alias{odp658c}
\title{deMenocal et al. ODP 658C Marine Sedimeent record}
\format{
A LiPD-TS-tibble with 8 LiPD variable columns
}
\source{
\url{https://lipdverse.org/HoloceneAbruptChange/0_10_0/ODP658C.deMenocal.1997.html}
}
\usage{
odp658c
}
\description{
deMenocal et al. ODP 658C Marine Sedimeent record
}
\keyword{datasets}
