% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{palmyra}
\alias{palmyra}
\title{Cobb et al. Palmyra coral d18O record}
\format{
A LiPD-TS-tibble with 4 LiPD variable columns
}
\source{
\url{https://lipdverse.org/Pages2kTemperature/2_1_0/Ocn-Palmyra.Cobb.2013.html}
}
\usage{
palmyra
}
\description{
Cobb et al. Palmyra coral d18O record
}
\keyword{datasets}
