% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plottings.R
\name{plot.excursion}
\alias{plot.excursion}
\title{Plot an excursion with uncertainties}
\usage{
\method{plot}{excursion}(x, ...)
}
\arguments{
\item{x}{The output of detectExcursion()}

\item{...}{additional parameters (see plotExcursion)}
}
\value{
a ggplot2 object
}
\description{
Plot an excursion with uncertainties
}
