% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plottings.R
\name{plot.shiftCore}
\alias{plot.shiftCore}
\title{Plot mean or variance shifts, with uncertainties and null hypothesis testing}
\usage{
\method{plot}{shiftCore}(x, ...)
}
\arguments{
\item{x}{Output from actR::detectShift}

\item{...}{additional inputs (see plotShiftCore)}
}
\value{
A ggplot object
}
\description{
Plot mean or variance shifts, with uncertainties and null hypothesis testing
}
