% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial.R
\name{plotDataDensity}
\alias{plotDataDensity}
\title{Data Coverage Plot}
\usage{
plotDataDensity(distance.grid = NULL, color.breaks = c(1, 2, 4, 8, 16))
}
\arguments{
\item{distance.grid}{as produced by distanceGrid()}

\item{color.breaks}{vector of breaks for coloring plot}
}
\value{
plot data
}
\description{
Data Coverage Plot
}
