% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotECDF.R
\name{plotECDF}
\alias{plotECDF}
\title{Plot ECDF (empirical cumulative distribution function) of edr10_obs and edr10_act}
\usage{
plotECDF(plot_position.df)
}
\arguments{
\item{plot_position.df}{A Data frame containing the plotting position}
}
\value{
A ggplot object
}
\description{
Plot ECDF (empirical cumulative distribution function) of edr10_obs and edr10_act
}
\examples{
plot_ecdf <- plotECDF(pp)
}

