% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plottings.R
\name{plotExcursion}
\alias{plotExcursion}
\title{Plot an excursion with uncertainties}
\usage{
plotExcursion(
  x,
  alpha = 0.05,
  print.significance = FALSE,
  x.axis.label = NA,
  y.axis.label = NA,
  lab.mult = 0.02
)
}
\arguments{
\item{x}{The output of detectExcursion()}

\item{alpha}{What significance level to use?}

\item{print.significance}{Show significance on the plot?}

\item{x.axis.label}{Label the x-axis (default = NA, which will automatically generate from input)}

\item{y.axis.label}{Label the y-axis (default = NA, which will automatically generate from input)}

\item{lab.mult}{When labeling significance, how much higher than the highest point to put the label}
}
\value{
a ggplot2 object
}
\description{
Plot an excursion with uncertainties
}
