% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plottings.R
\name{plotExcursionCore}
\alias{plotExcursionCore}
\title{Plot a basic excursion without propagated uncertainty}
\usage{
plotExcursionCore(exc.out.core, add.to.plot = ggplot2::ggplot())
}
\arguments{
\item{exc.out.core}{the output of detectExcursionCore()}

\item{add.to.plot}{a ggplot object upon which to add this plot}
}
\value{
a ggplot object
}
\description{
Plot a basic excursion without propagated uncertainty
}
