% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plottings.R
\name{plotExcursionSliding}
\alias{plotExcursionSliding}
\title{Plot a timeseries result of sliding}
\usage{
plotExcursionSliding(x, xlim = c(12000, 0))
}
\arguments{
\item{x}{sliding excursion output}
}
\value{
a plot object
}
\description{
Plot a timeseries result of sliding
}
