% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotLN.R
\name{plotLN}
\alias{plotLN}
\title{Plot Log normal cumulative distribution function) of edr10_obs act and edr10_act}
\usage{
plotLN(edr10_all)
}
\arguments{
\item{edr10_all}{A Data frame containing edr10_act and edr10_act, with their plotting position}

\item{PC_act}{A Data frame containing the parameters of edr10_act}

\item{PC_act}{A Data frame containing the parameters of edr10_act}
}
\value{
A ggplot object
}
\description{
Plot Log normal cumulative distribution function) of edr10_obs act and edr10_act
}
\examples{
plot_ln <- plotLN(edr10_all,PC_act, PC_gen)
}

