% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plottings.R
\name{plotNullHistogram}
\alias{plotNullHistogram}
\title{Plot a histogram of null hypothesis and KDE results}
\usage{
plotNullHistogram(x, pal = "Paired", h = NA)
}
\arguments{
\item{x}{a actR output object}

\item{pal}{Color brewer palette}

\item{h}{KDE smoothing factor, lower is smoother}
}
\value{
a ggplot object
}
\description{
Plot a histogram of null hypothesis and KDE results
}
