% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plottings.R
\name{plotSectionMeans}
\alias{plotSectionMeans}
\title{Plot a shift in mean or variance}
\usage{
plotSectionMeans(
  x,
  time,
  vals,
  add.to.plot = ggplot2::ggplot(),
  mean.color = "red"
)
}
\arguments{
\item{x}{Output of detectShiftCore()}

\item{time}{Input time vector}

\item{vals}{input value vector}

\item{add.to.plot}{ggplot to add this to}

\item{mean.color}{color of the mean lines}
}
\value{
A ggplot object
}
\description{
Plot a shift in mean or variance
}
