% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plottings.R
\name{plotShiftCore}
\alias{plotShiftCore}
\title{Plot a shift in mean or variance}
\usage{
plotShiftCore(ms.core.out, time, vals, add.to.plot = ggplot2::ggplot())
}
\arguments{
\item{ms.core.out}{Output of detectShiftCore()}

\item{time}{Input time vector}

\item{vals}{input value vector}

\item{add.to.plot}{ggplot to add this to}
}
\value{
A ggplot object
}
\description{
Plot a shift in mean or variance
}
