% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ploting_position.R
\name{plotting_position}
\alias{plotting_position}
\title{Calculating plotting position of edr10 obs and edr_10 act
Plooting position are calculated according method described in SSWD help doc page 15}
\usage{
plotting_position(Edr10)
}
\arguments{
\item{Edr10}{A dataframe containing edr10}
}
\value{
the same dataframe with added columns containing the ploting position (w_i_j) and intermediate variables
}
\description{
Calculating plotting position of edr10 obs and edr_10 act
Plooting position are calculated according method described in SSWD help doc page 15
}
\examples{
pp_obs <- plotting_position (edr10_obs)
}

