% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict_PC.R
\name{predict_PC}
\alias{predict_PC}
\title{Prediction of muC_reg and SigmaC_reg}
\usage{
predict_PC(PAC_obs, form_mu, form_sigma)
}
\arguments{
\item{PAC_obs}{dataframe containing the observed  acute and chronic weighted parameters}

\item{form_mu}{Formula of the best model selected to predict muC_act}

\item{form_sigma}{Formula of the best model selected to predict SigmaC_act}
}
\value{
a list with the Chronic predicted  parameters (MuC_act and SigmaC_act) ,
regression coefficients for MuC_reg and SigmaC_reg and sd of residuals for  both regressions
NB : means of residuals = 0 by definition
}
\description{
The function takes in argument the observed weigted parameters of ED50
i.e mu_A and sigma_A, and the formula of the best models for prediction
of mu_C_act and sigma_C_act and return the muC_act and SigmaC_act
They are estimated using a cross validation. When parameters of one class are
predicted data of that class are not taken in acount in the regression step
}
\examples{
library(ACTpckg)
data(cipr)
ciprKP <- subset_data(cipr, Class,6)
PAC_obs<-est_PAC(ciprKP,Class)
PC_pred <- predict_PC(PAC_obs,bm_mu, bm_sigma)

}

