% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uq.R
\name{propagateUncertainty}
\alias{propagateUncertainty}
\title{Detect a changepoint while propagating proxy and/or age uncertainty}
\usage{
propagateUncertainty(
  age,
  vals,
  changeFun,
  simulateAgeUncertainty = TRUE,
  simulatePaleoUncertainty = TRUE,
  n.ens = 100,
  bam.model = list(ns = n.ens, name = "bernoulli", param = 0.05),
  paleo.uncertainty = sd(vals, na.rm = TRUE),
  paleo.ar1 = sqrt(0.5),
  paleo.arima.order = c(1, 0, 0),
  summarize = FALSE,
  ...
)
}
\arguments{
\item{...}{}
}
\value{

}
\description{
Detect a changepoint while propagating proxy and/or age uncertainty
}
