% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{pseudoShifts}
\alias{pseudoShifts}
\title{A synthetic example of shifts in mean statee for testing and illustration, with and without age ensembles}
\format{
A LiPD-TS-tibble with 4 LiPD variable columns, age and temperature ensembles
}
\usage{
pseudoShifts
}
\description{
A synthetic example of shifts in mean statee for testing and illustration, with and without age ensembles
}
\author{
Julien Emile-Geay
}
\keyword{datasets}
