% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reexports.R
\docType{import}
\name{reexports}
\alias{reexports}
\alias{surrogateDataFun}
\alias{simulateAutoCorrelatedUncertainty}
\alias{propagateUncertainty}
\alias{testNullHypothesis}
\alias{kdePval}
\title{Objects exported from other packages}
\keyword{internal}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{ens}{\code{\link[ens]{kdePval}}, \code{\link[ens]{propagateUncertainty}}, \code{\link[ens]{simulateAutoCorrelatedUncertainty}}, \code{\link[ens]{surrogateDataFun}}, \code{\link[ens]{testNullHypothesis}}}
}}

