% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resume_hdr.R
\name{resume_hdr}
\alias{resume_hdr}
\title{Print estimate of hdrX_obs and hdrX_act, and their ratio (obs /act) and the mean and median of this ratio}
\usage{
resume_hdr(myHdrX_obs, myHdrX_act)
}
\arguments{
\item{myHdr_obs}{Data frame containing hdrX_obs}

\item{myHdr_act}{Data frame containing hdrX_obs}
}
\value{
a list containing hdrX_obs and hdrX_act, and their ration (obs /act) and the mean and median of this ratio
}
\description{
Print estimate of hdrX_obs and hdrX_act, and their ratio (obs /act) and the mean and median of this ratio
}
\examples{
myhdr5 <-resume_hdr(hdr5_obs, hdr5_act)
}

