% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rm_outliers.R
\name{rm_outliers}
\alias{rm_outliers}
\title{Return a dataset without the identified outliers according coefficient
This function is based on boxplot.stat function}
\usage{
rm_outliers(data, coef)
}
\arguments{
\item{data}{A dataframe containing the edr10 or ed50}

\item{coef}{this determines how far the plot ‘whiskers’ extend out from the
box. If coef is positive, the whiskers extend to the most extreme data point
which is no more than coef times the length of the box away from the box.
A value of zero causes the whiskers to extend to the data extremes (and no outliers be returned).}
}
\value{
the dataset without the outliers
}
\description{
Return a dataset without the identified outliers according coefficient
This function is based on boxplot.stat function
}
\examples{
ciprKP1_5.Ac<-rm_outliers(ciprKP0.Ac,1.5)
}

