% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sd_res_reg.R
\name{sd_res_reg}
\alias{sd_res_reg}
\title{Return standard deviation of residuals  of the best models (1 for Mu, 1 for Sigma) applied on observed data}
\usage{
sd_res_reg(PAC_obs, form_mu, form_sigma)
}
\arguments{
\item{PAC_obs}{dataframe containing the observed  acute and chronic weighted parameters}

\item{form_mu}{Formula of the best model selected to predict muC_act}

\item{form_sigma}{Formula of the best model selected to predict SigmaC_act}
}
\value{
a list with the residuals standrd deviation of the best model applied on observed data
}
\description{
Return standard deviation of residuals  of the best models (1 for Mu, 1 for Sigma) applied on observed data
}
\examples{
library(ACTR)
data(cipr)
ciprKP <- subset_data(cipr, Class,6)
PAC_obs <- est_PAC(ciprKP,Class)
sd_res <- sd_res_reg (PAC_obs,bm_mu, bm_sigma)
}

