% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/show_outliers.R
\name{show_outliers}
\alias{show_outliers}
\title{Returns outliers with var_group, ID, sub_ID and Ed values}
\usage{
show_outliers(data, var_grp, coef)
}
\arguments{
\item{var_grp}{A group variable telling how outliers have to be detected}

\item{coef}{this determines how far the plot ‘whiskers’ extend out from the
box. If coef is positive, the whiskers extend to the most extreme data point
which is no more than coef times the length of the box away from the box.
A value of zero causes the whiskers to extend to the data extremes (and no outliers be returned).}

\item{dataset}{A dataframe}
}
\value{
the detected Outliers by var_grp level
}
\description{
Returns outliers with var_group, ID, sub_ID and Ed values
}
\examples{

}

