% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/split_edr10_obs.R
\name{split_edr10_obs}
\alias{split_edr10_obs}
\title{Split edr10_obs from dataset}
\usage{
split_edr10_obs(dat, var_grp)
}
\arguments{
\item{dat}{A dataframe containing edr10 and ed50}

\item{var_grp}{A variable group (kept in the return dataset, i.e. Class)}
}
\value{
a data frame contianing the edr10_obs, var_grp (i.e. Class or Phylum), SpeciesComp, Id, subid and Umbrella effect
}
\description{
Split edr10_obs from dataset
}
\examples{
edr10_obs <- split_edr10_obs(ciprKP, Class)
}

