% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subset_data.R
\name{subset_data}
\alias{subset_data}
\title{Subset dataset according a minimal number of ED desired by modalities of group variable (ie. Class)}
\usage{
subset_data(dat, var_grp, n_min)
}
\arguments{
\item{var_grp}{A group variable telling how counts have to be done (ie.by Class)}

\item{n_min}{A integer telling the minimum number of EDR10 and ED50 desired}

\item{dataset}{A dataframe}
}
\value{
a subset of the original dataset
}
\description{
Subset dataset according a minimal number of ED desired by modalities of group variable (ie. Class)
}
\examples{
data(cipr)
new<-subset_data(cipr, Class, 6)
new<-subset_data(cipr, Phylum, 6)
}

