% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarize.R
\name{summarizeChangepointMeanAndSd}
\alias{summarizeChangepointMeanAndSd}
\title{Summarize changepoint mean and standard deviation}
\usage{
summarizeChangepointMeanAndSd(x, alpha = 0.05)
}
\arguments{
\item{x}{Output of detectShiftCore()}

\item{alpha}{significance level of changepoints to consider}
}
\value{
a tibble of change point mean changes
}
\description{
Summarize changepoint mean and standard deviation
}
