% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarize.R
\name{summarizeEventProbability}
\alias{summarizeEventProbability}
\title{Summarize the shift detection results using a uniform bin step}
\usage{
summarizeEventProbability(
  exc.out,
  bin.vec = NA,
  bin.step = 10,
  max.time = NA,
  min.time = NA,
  shift.type = "either",
  time.dir = "retrograde"
)
}
\arguments{
\item{exc.out}{shiftCore (tbl_df) class result of propagateUncertainty()}

\item{bin.vec}{Vector by which to bin the data}

\item{bin.step}{Number of years to group shifts by}

\item{max.time}{Maximum age of bin.vec to summarize}

\item{min.time}{Minimum age of bin.vec to summarize}

\item{shift.type}{Type of excursion to look for. "positive", "negative", "either" or "both" (default = "either")}

\item{time.dir}{what way does time go, "retrograde" (default) (backwards, e.g. yr BP units) or "prograde" (forward e.g AD units)}
}
\value{
a tibble of event probability for each bin step
}
\description{
Summarize the shift detection results using a uniform bin step
}
