% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.R
\name{summary.excursion}
\alias{summary.excursion}
\title{Summarize Excursion Output}
\usage{
\method{summary}{excursion}(object, ...)
}
\arguments{
\item{object}{excursion class output}

\item{...}{additional inputs}
}
\description{
Summarize Excursion Output
}
