% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.R
\name{summary.shift}
\alias{summary.shift}
\title{Summarize shift output}
\usage{
\method{summary}{shift}(object, ...)
}
\arguments{
\item{object}{shift output}

\item{...}{additional inputs (see summaryShift)}
}
\value{

}
\description{
Summarize shift output
}
