% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.R
\name{summaryExcursion}
\alias{summaryExcursion}
\title{Summarize Excursion Output}
\usage{
summaryExcursion(
  object,
  params.to.print = c("sig.num", "n.consecutive", "exc.type", "min.vals", "na.rm")
)
}
\arguments{
\item{object}{excursion class output}

\item{params.to.print}{a vector of parameters to print}
}
\description{
Summarize Excursion Output
}
