% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.R
\name{summaryShift}
\alias{summaryShift}
\title{Summarize shift output}
\usage{
summaryShift(
  object,
  alpha = 0.05,
  params.to.print = c("cpt.fun", "minimum.segment.length", "method", "penalty",
    "ncpts.max")
)
}
\arguments{
\item{object}{shift output}

\item{alpha}{significance level}

\item{params.to.print}{vector of paramters to print}
}
\description{
Summarize shift output
}
