% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uq.R
\name{surrogateDataFun}
\alias{surrogateDataFun}
\title{Use the write function from rEDM depending on the version}
\usage{
surrogateDataFun(...)
}
\arguments{
\item{...}{Arguments to pass to surrogate}
}
\value{
Surrogate output
}
\description{
Use the write function from rEDM depending on the version
}
