% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validCountries.R
\name{validCountries}
\alias{validCountries}
\title{Valid Country Codes of HS Lookups}
\usage{
validCountries()
}
\value{
Returns a named character vector.  The names are the valid country
codes that can be passed to the ACT API and the values are the common nmaes.
}
\description{
Function to return all valid country codes used by the ACT API
}
\examples{
validCountries()
}
