% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validStates.R
\name{validStates}
\alias{validStates}
\title{Valid State Codes of HS Lookups}
\usage{
validStates()
}
\value{
Returns a named character vector.  The names are the valid state
codes that can be passed to the ACT API and the values are the common nmaes.
}
\description{
Function to return all valid state codes used by the ACT API
}
