% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/view_hdr.R
\name{view_hdr}
\alias{view_hdr}
\title{Plot hdr_th_obs and hdr_th_act with their 95 percent confidence intervals}
\usage{
view_hdr(myHdrX_obs, myHdrX_act)
}
\arguments{
\item{myHdr5_obs}{Data frame containing hdr5_obs}

\item{myHdr5_act}{Data frame containing hdr5_obs}
}
\value{
A ggplot object
}
\description{
Plot hdr_th_obs and hdr_th_act with their 95 percent confidence intervals
}
\examples{
plot_hdr5 <-view_hdr(hdr5_obs, hdr5_act)
}

