% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/excursion.R
\name{detectExcursion}
\alias{detectExcursion}
\title{Detect an excursion in a timeseries}
\usage{
detectExcursion(
  age,
  vals,
  event.yr,
  event.window,
  ref.window,
  n.ens = 100,
  output.figure.path = NA,
  ...
)
}
\arguments{
\item{age}{a time vector, or matrix of time ensemble members (ensembles in columns)}

\item{vals}{a values vector, or matrix of values ensemble members (ensembles in columns)}

\item{event.yr}{the center of the proposed excursion window}

\item{event.window}{the width of the proposed excursion window}

\item{ref.window}{how many years to use as a reference before and after the event window}

\item{n.ens}{How many ensembles to propagate through? (default = 100)}

\item{output.figure.path}{path pointing to where should the output figure be saved? An NA will not produce a figure (default = NA)}

\item{...}{
  Arguments passed on to \code{\link[=propagateUncertainty]{propagateUncertainty}}
  \describe{
    \item{\code{changeFun}}{the change function to across which to propagate}
    \item{\code{simulateAgeUncertainty}}{TRUE or FALSE. If an ensemble is not included, do you want to simulate age ensembles (default = TRUE)}
    \item{\code{simulatePaleoUncertainty}}{TRUE or FALSE. If an ensemble is not included, do you want to simulate paleo ensembles (default = TRUE)}
    \item{\code{bam.model}}{BAM Model parameters to use if simulating age uncertainty (default = list(ns = n.ens, name = "bernoulli", param = 0.05), paleo.uncertainty = sd(vals,na.rm = TRUE)))}
    \item{\code{paleo.uncertainty}}{Uncertainty to use if modelling uncertainty for paleo values. (default = sd(vals,na.rm = TRUE)/2)}
    \item{\code{paleo.ar1}}{Autocorrelation coefficient to use for modelling uncertainty on paleoData, what fraction of the uncertainties are autocorrelated? (default = sqrt(0.5); or 50\% autocorrelated uncertainty)}
    \item{\code{paleo.arima.order}}{Order to use for ARIMA model used in modelling uncertainty on paleoDat (default = c(1,0,0))}
    \item{\code{summarize}}{Boolean. Summarize the output? Or return all the ensembles?}
  }}
}
\value{
a tibble that describes the positive and negative excursion results
}
\description{
Determines whether an excursion event has occurred within the specified event window. Excursion events are defined as n.consecutive values within the event window that are more extreme than the avg +/- sig.num standard deviations of the reference windows.
}
\references{
Morrill
}
\author{
Hannah Kolus

Nick McKay
}
