% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/excursion.R
\name{detectExcursion}
\alias{detectExcursion}
\title{Detect an excursion in a timeseries}
\usage{
detectExcursion(
  age,
  vals,
  event.yr,
  event.window,
  ref.window,
  n.consecutive = 2,
  output.figure.path = NA,
  sig.num = 2
)
}
\arguments{
\item{age}{a year or age vector}

\item{vals}{a vector of values the same length as age}

\item{event.yr}{the center of the proposed excursion window}

\item{event.window}{the width of the proposed excursion window}

\item{ref.window}{how many years to use as a reference before and after the event window}

\item{n.consecutive}{how many consecutive points are required for this to be considered an excursion? (default = 2)}

\item{output.figure.path}{path pointing to where should the output figure be saved? An NA will not produce a figure (default = NA)}

\item{sig.num}{how many standard deviations required outside the reference windows must be exceeded for this to be considered an excursion? (default = 2)}
}
\value{
a tibble that describes the positive and negative excursion results
}
\description{
Determines whether an excursion event has occurred within the specified event window. Excursion events are defined as n.consecutive values within the event window that are more extreme than the avg +/- sig.num standard deviations of the reference windows.
}
\references{
Morrill
}
\author{
Hannah Kolus

Nick McKay
}
