% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/excursion.R
\name{detectExcursionCore}
\alias{detectExcursionCore}
\title{Detect excursion - core functionality}
\usage{
detectExcursionCore(
  time,
  vals,
  event.yr,
  event.window,
  ref.window,
  sig.num = 2,
  n.consecutive = 2,
  exc.type = "either",
  min.vals = 4,
  na.rm = TRUE
)
}
\arguments{
\item{time}{time vector of only the points in the window}

\item{vals}{value vector of only the points in the window}

\item{event.yr}{time at the center of the excursion window}

\item{event.window}{width (in time units) of the excursion window}

\item{ref.window}{width (in time units) of the reference windows}

\item{sig.num}{how many standard deviations required outside the reference windows must be exceeded for this to be considered an excursion? (default = 2)}

\item{n.consecutive}{how many consecutive points are required for this to be considered an excursion? (default = 2)}

\item{exc.type}{Type of excursion to look for. "positive", "negative", "either" or "both" (default = "either")}

\item{min.vals}{Minimum effective sample size (adjusted by autocorrelation) required in reference and event windows (default = 4)}

\item{na.rm}{Remove NAs? (default = TRUE)}
}
\value{
a tibble of results
}
\description{
Detect excursion - core functionality
}
