% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/excursion.R
\name{detectExcursionCore}
\alias{detectExcursionCore}
\title{Detect excursion - core functionality}
\usage{
detectExcursionCore(
  age,
  vals,
  event.start,
  event.end,
  sig.num = 2,
  n.consecutive = 2,
  exc.type = "either",
  min.vals = 8,
  na.rm = TRUE
)
}
\arguments{
\item{age}{age vector of only the points in the window}

\item{vals}{value vector of only the points in the window}

\item{event.start}{the start of the event window}

\item{event.end}{the end of the event window}

\item{sig.num}{how many standard deviations required outside the reference windows must be exceeded for this to be considered an excursion? (default = 2)}

\item{n.consecutive}{how many consecutive points are required for this to be considered an excursion? (default = 2)}
}
\value{
a tibble of results
}
\description{
Detect excursion - core functionality
}
