% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meanShift.R
\name{detectShift}
\alias{detectShift}
\title{Detect a shift in the mean and/or variance of a dataset}
\usage{
detectShift(
  ltt = NA,
  time = NA,
  vals = NA,
  time.variable.name = NA,
  vals.variable.name = NA,
  time.units = NA,
  vals.units = NA,
  dataset.name = NA,
  surrogate.method = "isospectral",
  summary.bin.step = 100,
  summary.bin.vec = NA,
  null.hypothesis.n = 100,
  null.quantiles = c(0.95, 0.9),
  time.range = NA,
  calc.deltas = TRUE,
  ...
)
}
\arguments{
\item{ltt}{A LiPD-timeseries-tibble, a tibble or data.frame that has the variable(s) of interest, a time variable (age, year or time) along with their metadata, aranged in rows. If ltt = NA, then one in is created from other inputs}

\item{time}{if ltt is not provided, input a vector or matrix of time (year or age) data. If it's a multicolumn matrix, the columns are time-ensemble members}

\item{vals}{if ltt is not provided, input a vector or matrix of paleoData. If it's a multicolumn matrix, the columns are value-ensemble members}

\item{time.variable.name}{If ltt is not provided, specify the name of the time variable (typically 'age' or 'year')}

\item{vals.variable.name}{If ltt is not provided, specify the name of the paleo variable (e.g., 'd18O' or 'temperature'). Alternatively, if ltt is provided with more rows than expected, this term is used to attempt to select the correct row.}

\item{time.units}{If ltt is not provided, specify the units the time variable (typically 'yr BP' or 'CE')}

\item{vals.units}{If ltt is not provided, specify the units the paleo variable (e.g. 'permil' or 'degrees C')}

\item{dataset.name}{If ltt is not provided, specify the dataset name}

\item{surrogate.method}{What method to use to generate surrogate data. Options:
\itemize{
\item 'isospectral': (Default) Following Ebisuzaki (1997), generate surrogates by scrambling the phases of the data while preserving their power spectrum. Uses surrogateDataFun() (rEDM::SurrogateData).
\item 'isopersistent': Generates surrogates by simulating from an AR(1) process fit to the data. Uses createSyntheticTimeseries().
\item 'shuffle': Randomly shuffles the data. Uses surrogateDataFun().
}}

\item{summary.bin.step}{Time interval over which to summarize the results}

\item{summary.bin.vec}{Optionally provide a vector over which to create the summary bins, this will supersede summary.bin.step if provided (default = NA)}

\item{null.hypothesis.n}{How many simulations to run for null hypothesis testing (default = 100)}

\item{null.quantiles}{What quantiles to report as output from null hypothesis testing (default = c(.95, .9))}

\item{time.range}{Optionally enter a time range (as minimum and maximum) that you'd like to restrict the analysis to. (default = NA)}

\item{calc.deltas}{Calculate the difference in means between change point sections. Default FALSE}

\item{...}{arguments to pass to changeFun. Vectors longer than 1 are treated as parameter ensembles and sampled across the n.ens members.}
}
\value{
A tibble of output data and metadata. Each row represents a time step of the summary bin
}
\description{
detectShift() allows you to detect a shift in the mean and/or variance of a dataset, and assess its significance given age and data uncertainty relative to a robust null hypothesis. This approach uses the function changepoint::cpt.mean(), changepoint::cpt.var(), or changepoint::cpt.meanvar()  from the changepoint package, propagates inputted or modelled time and/or value ensembles, and summarizes their likelihoods relative to a robust null hypothesis (see ?testNullHypothesis)
}
