% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/excursion.R
\name{maxConsecutive}
\alias{maxConsecutive}
\title{Maximum number of consecutive values in a vector}
\usage{
maxConsecutive(vec, val = TRUE)
}
\arguments{
\item{vec}{a vector}

\item{val}{the value to search for consecutive values}
}
\value{
the number of consecutive values. 0 means that value is not present in the vector
}
\description{
Maximum number of consecutive values in a vector
}
\examples{
maxConsecutive(c(1,2,7,5,7,7,7,3,4),val = 7)
}
\concept{https://stackoverflow.com/questions/37345853/find-the-maximum-and-mean-length-of-the-consecutive-true-arguments}
