% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plottings.R
\name{plot.excursion}
\alias{plot.excursion}
\title{Plot an excursion with uncertainties}
\usage{
plot.excursion(x, alpha = 0.05, print.significance = FALSE, lab.mult = 0.02)
}
\arguments{
\item{x}{The output of detectExcursion()}

\item{alpha}{What significance level to use?}

\item{print.significance}{Show significance on the plot?}

\item{lab.mult}{When labeling significance, how much higher than the highest point to put the label}
}
\value{
a ggplot2 object
}
\description{
Plot an excursion with uncertainties
}
