% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plottings.R
\name{plot.shift}
\alias{plot.shift}
\title{Plot mean or variance shifts, with uncertainties and null hypothesis testing}
\usage{
\method{plot}{shift}(x, ...)
}
\arguments{
\item{x}{Output from actR::detectShift}

\item{...}{more inputs, see (plotShift)}
}
\value{
a ggplot object
}
\description{
Plot mean or variance shifts, with uncertainties and null hypothesis testing
}
