% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plottings.R
\name{plot.shiftCore}
\alias{plot.shiftCore}
\title{Plot a shift in mean or variance}
\usage{
plot.shiftCore(x, line.color = "black", mean.color = "red")
}
\arguments{
\item{x}{tibble with time and vals}

\item{line.color}{color of the line}

\item{mean.color}{color of the means}
}
\value{
A ggplot object
}
\description{
Plot a shift in mean or variance
}
