% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plottings.R
\name{plotExcursion}
\alias{plotExcursion}
\title{Plot an excursion with uncertainties}
\usage{
plotExcursion(exc.out)
}
\arguments{
\item{exc.out}{The output of detectExcursion()}
}
\value{
a ggplot2 object
}
\description{
Plot an excursion with uncertainties
}
