% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plottings.R
\name{plotExcursionCore}
\alias{plotExcursionCore}
\title{Plot a basic excursion without uncertainty}
\usage{
plotExcursionCore(exc.out.core, add.to.plot = ggplot2::ggplot())
}
\arguments{
\item{add.to.plot}{a ggplot object upon which to add this plot}

\item{exc.out}{A 1 row tibble describing the excursion - this the output of detectExcursionCore, or part of the output of detectExcursion}
}
\value{

}
\description{
Plot a basic excursion without uncertainty
}
