% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plottings.R
\name{plotShift}
\alias{plotShift}
\title{Plot mean or variance shifts, with uncertainties and null hypothesis testing}
\usage{
plotShift(
  ms.out,
  cl.color = "Reds",
  plot.sig.vlines = TRUE,
  label.sig = TRUE,
  alpha = 0.05
)
}
\arguments{
\item{ms.out}{Output from actR::detectShift}

\item{cl.color}{Color palette, single color, or vector of colors to use for confidence intervals (default = "Reds"s)}

\item{plot.sig.vlines}{Plot vertical lines at significant change points? (default = TRUE)}

\item{label.sig}{Label significant change points (default = TRUE)}

\item{alpha}{significance level (default = 0.05)}
}
\value{

}
\description{
Plot mean or variance shifts, with uncertainties and null hypothesis testing
}
