% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial.R
\name{plotSignificance}
\alias{plotSignificance}
\title{plot results of spatialSigTest()}
\usage{
plotSignificance(
  pval.grid = NULL,
  sigTestResults,
  color.breaks = c(0.001, 0.01, 0.05, 0.1, 0.2),
  which.test = "pval",
  restrict.sites = TRUE,
  alpha.by.weight = TRUE,
  cutoff.distance = 1500,
  x.lim = c(-180, 180),
  y.lim = c(-90, 90),
  projection =
    "+proj=robin +lon_0=0 +x_0=0 +y_0=0 +ellps=WGS84 +datum=WGS84 +units=m +no_defs"
)
}
\arguments{
\item{pval.grid}{from spatialSigTest()}

\item{sigTestResults}{from excursionTestHighRes()}

\item{color.breaks}{at what significance levels should we put the color breaks?}
}
\value{
plot and plot data
}
\description{
plot results of spatialSigTest()
}
