% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{print.excursion}
\alias{print.excursion}
\title{Print Excursion}
\usage{
\method{print}{excursion}(x, ...)
}
\arguments{
\item{x}{excursion object to print}

\item{...}{
  Arguments passed on to \code{\link[=print]{print}}
  \describe{
    \item{\code{}}{}
  }}
}
\description{
Print Excursion
}
