% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{print.shift}
\alias{print.shift}
\title{Print shift}
\usage{
\method{print}{shift}(x, ...)
}
\arguments{
\item{x}{shift object to print}

\item{...}{
  Arguments passed on to \code{\link[=summary.shift]{summary.shift}}
  \describe{
    \item{\code{object}}{shift output}
  }}
}
\description{
Print shift
}
