% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial.R
\name{spatialSigTest}
\alias{spatialSigTest}
\title{aggregate excursion tests spatially}
\usage{
spatialSigTest(
  map.grid.cell,
  events,
  agg.method = "robustNull",
  min.pval = 0.001,
  distance.cutoff = 2000,
  use.weights = TRUE
)
}
\arguments{
\item{map.grid.cell}{from distanceGrid()}

\item{events}{an events object}

\item{agg.method}{What method do you want to use to aggregate the p-values? Choose from robustNull (default), fisher, sidak or lancaster}

\item{min.pval}{min p value}
}
\value{
gridded significance results
}
\description{
aggregate excursion tests spatially
}
