% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uq.R
\name{testNullHypothesis}
\alias{testNullHypothesis}
\title{Detect excursions in synthetic datasets that mimic a real on}
\usage{
testNullHypothesis(
  age,
  vals,
  changeFun,
  method = "isospectral",
  n.ens = 100,
  ...
)
}
\arguments{
\item{age}{a year or age vector}

\item{vals}{a vector of values the same length as age}

\item{n.ens}{The number of ensembles to simulate (default = 100)}
}
\value{
a tibble that reports the positivity rate in the synthetics
}
\description{
Detect excursions in synthetic datasets that mimic a real on
}
